refactor(useDaemon): extract useChat into its own file, fix 300-line limit
